# API.yaml
...
swagger: "2.0"
info:
  description: "Calculates minimum path between points depending on user generated\
    \ map input."
  version: "1.0.0"
  title: "Team 2 Pathfinding API"
host: "ec2-54-200-185-55.us-west-2.compute.amazonaws.com:8080"
basePath: "/v2"
schemes:
- "http"
paths:
  /:
    post:
      tags:
      - "Problems"
      summary: "Creates a new problem and returns a problemID"
      description: ""
      operationId: "add_problem"
      produces:
      - "application/json"
      - "application/xml"
      parameters: []
      responses:
        200:
          description: "The id of the new problem generated"
          schema:
            type: "integer"
            format: "int32"
        405:
          description: "Invalid input"
      x-swagger-router-controller: "swagger_server.controllers.problems_controller"
  /id={problem_id}/:
    get:
      tags:
      - "Problems"
      summary: "Problems"
      description: "Returns a specific problem\n"
      operationId: "get_problem"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "problem_id"
        in: "path"
        description: "The id of the problem being manipulated"
        required: true
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "A Problem type, containing all values"
          schema:
            $ref: "#/definitions/Problem"
        403:
          description: "ProblemID busy"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "swagger_server.controllers.problems_controller"
    delete:
      tags:
      - "Problems"
      summary: "Delete Problem"
      description: "This removes the problem by the given ID"
      operationId: "delete_problem"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "problem_id"
        in: "path"
        description: "The id of the problem being manipulated"
        required: true
        type: "integer"
        format: "int32"
      responses:
        400:
          description: "Invalid ID supplied"
        403:
          description: "ProblemID busy"
        404:
          description: "Problem not found"
      x-swagger-router-controller: "swagger_server.controllers.problems_controller"
...
  /id={problem_id}/Robot:
    get:
      tags:
      - "Robot"
      summary: "Robot"
      description: "Returns a description of the robots, including the current location\n"
      operationId: "get_robots"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "problem_id"
        in: "path"
        description: "The id of the problem being manipulated"
        required: true
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "A robot object array, including current location."
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Robot"
        403:
          description: "ProblemID busy"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "swagger_server.controllers.robot_controller"
    post:
      tags:
      - "Robot"
      summary: "Add a new robot to the list"
      description: ""
      operationId: "add_robot"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "problem_id"
        in: "path"
        description: "The id of the problem being manipulated"
        required: true
        type: "integer"
        format: "int32"
      - in: "body"
        name: "robot"
        description: "Obstacle object that needs to be added to the list."
        required: true
        schema:
          $ref: "#/definitions/Robot"
      responses:
        200:
          description: "The id of the new robot generated"
          schema:
            type: "integer"
            format: "int32"
        403:
          description: "ProblemID busy"
        405:
          description: "Invalid input"
      x-swagger-router-controller: "swagger_server.controllers.robot_controller"
  /id={problem_id}/Robot/rid={robot_id}:
    get:
      tags:
      - "Robot"
      summary: "Get a robot by the ID"
      description: ""
      operationId: "get_robot"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "problem_id"
        in: "path"
        description: "The id of the problem being manipulated"
        required: true
        type: "integer"
        format: "int32"
      - name: "robot_id"
        in: "path"
        description: "Robot object that needs to be updated."
        required: true
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "A robot"
          schema:
            $ref: "#/definitions/Robot"
        403:
          description: "ProblemID busy"
        404:
          description: "Robot not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "swagger_server.controllers.robot_controller"
    put:
      tags:
      - "Robot"
      summary: "Update the existing robot value"
      description: ""
      operationId: "update_robot"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "problem_id"
        in: "path"
        description: "The id of the problem being manipulated"
        required: true
        type: "integer"
        format: "int32"
      - in: "body"
        name: "robot"
        description: "Robot object that needs to be updated."
        required: true
        schema:
          $ref: "#/definitions/Robot"
      - name: "robot_id"
        in: "path"
        description: "Robot object that needs to be updated."
        required: true
        type: "integer"
        format: "int32"
      responses:
        403:
          description: "ProblemID busy"
        404:
          description: "Robot not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "swagger_server.controllers.robot_controller"
    delete:
      tags:
      - "Robot"
      summary: "Delete Robot"
      description: "This removes the robot by the given ID"
      operationId: "delete_robot"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "problem_id"
        in: "path"
        description: "The id of the problem being manipulated"
        required: true
        type: "integer"
        format: "int32"
      - name: "robot_id"
        in: "path"
        description: "The ID of the Obstacle that needs to be deleted."
        required: true
        type: "integer"
        format: "int32"
      responses:
        400:
          description: "Invalid ID supplied"
        403:
          description: "ProblemID busy"
        404:
          description: "Robot not found"
      x-swagger-router-controller: "swagger_server.controllers.robot_controller"
...
definitions:
  Coordinate:
    type: "object"
    properties:
      longitude:
        type: "number"
        format: "double"
        description: "the longitude of the object."
      latitude:
        type: "number"
        format: "double"
        description: "the latitude of the object."
  Rectangle:
    type: "object"
    properties:
      coordinates:
        $ref: "#/definitions/Coordinate"
      width:
        type: "number"
        format: "double"
        description: "The width of the rectangle"
      height:
        type: "number"
        format: "double"
        description: "The height of the rectangle"
  Obstacle:
    type: "object"
    properties:
      obstacle_id:
        type: "integer"
        format: "int32"
        description: "Unique identifier representing a specific obstacle in the list."
      obstacle_info:
        $ref: "#/definitions/Rectangle"
  Robot:
    type: "object"
    properties:
      coordinates:
        $ref: "#/definitions/Coordinate"
      id:
        type: "integer"
        format: "int64"
  Goal:
    type: "object"
    properties:
      coordinates:
        $ref: "#/definitions/Coordinate"
  Boundary:
    type: "object"
    properties:
      boundary_info:
        $ref: "#/definitions/Rectangle"
...
  Problem:
    type: "object"
    properties:
      problem_id:
        type: "integer"
        format: "int32"
      robots:
        type: "array"
        description: "The list of robots for the problem"
        items:
          $ref: "#/definitions/Robot"
      goal:
        $ref: "#/definitions/Goal"
      boundary:
        $ref: "#/definitions/Boundary"
      obstacles:
        type: "array"
        items:
          $ref: "#/definitions/Obstacle"
  Error:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      message:
        type: "string"
      fields:
        type: "string"